requirements.txt
aiogram>=3.23.0
python-dotenv>=1.2.1
requests>=2.32.0
SQLAlchemy>=2.0.36
alembic>=1.13.3
